# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  completed: 'completed',
} as const;

export interface Sport {
  id: number;
  name: string;
}

export interface Player {
  id: number;
  sportId: number;
  name: string;
  country: string;
  /** @nullable */
  currentRanking?: number | null;
}

export interface Tournament {
  id: number;
  sportId: number;
  name: string;
  region: string;
  surface: string;
  category: string;
  season: number;
}

export interface MatchSet {
  id: number;
  matchId: number;
  setNumber: number;
  playerAGames: number;
  playerBGames: number;
}

export interface Match {
  id: number;
  sportId: number;
  tournamentId: number;
  tournamentName: string;
  playerA: Player;
  playerB: Player;
  date: string;
  status: MatchStatus;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  resultSummary: string | null;
  sets?: MatchSet[];
}

export interface HeadToHeadSurface {
  surface: string;
  totalMeetings: number;
  playerAWins: number;
  playerBWins: number;
  playerAWinPercentage: number;
  playerBWinPercentage: number;
}

export interface HeadToHeadMatch {
  id: number;
  date: string;
  tournamentName: string;
  surface: string;
  playerA: Player;
  playerB: Player;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  resultSummary: string | null;
}

export interface HeadToHead {
  totalMeetings: number;
  playerAWins: number;
  playerBWins: number;
  playerAWinPercentage: number;
  playerBWinPercentage: number;
  matchesBeforeScheduledTime: number;
  surfaceBreakdown: HeadToHeadSurface[];
  recentMeetings: HeadToHeadMatch[];
}

export type MatchDetail = Match & {
  tournament: Tournament;
  headToHead: HeadToHead;
};

export interface Dashboard {
  sport: Sport;
  today: Match[];
  live: Match[];
  upcoming: Match[];
  completed: Match[];
}

export interface PlayerDetail {
  player: Player;
  recentMatches: Match[];
  wins: number;
  losses: number;
  winPercentage: number;
}

export interface TournamentDetail {
  tournament: Tournament;
  upcomingMatches: Match[];
  completedMatches: Match[];
}

export interface SearchResults {
  players: Player[];
  tournaments: Tournament[];
  matches: Match[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type SportIdParameter = number;

export type TournamentIdParameter = number;

export type PlayerIdParameter = number;

export type GetDashboardParams = {
/**
 * @minimum 1
 */
sportId?: SportIdParameter;
};

export type ListMatchesParams = {
/**
 * @minimum 1
 */
sportId?: SportIdParameter;
/**
 * @minimum 1
 */
tournamentId?: TournamentIdParameter;
/**
 * @minimum 1
 */
playerId?: PlayerIdParameter;
status?: MatchStatus;
};

export type SearchParams = {
/**
 * @minLength 1
 */
q: string;
/**
 * @minimum 1
 */
sportId?: SportIdParameter;
};