# lib/api-zod/src/generated/types/headToHead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeadToHeadMatch } from './headToHeadMatch';
import type { HeadToHeadSurface } from './headToHeadSurface';

export interface HeadToHead {
  totalMeetings: number;
  playerAWins: number;
  playerBWins: number;
  playerAWinPercentage: number;
  playerBWinPercentage: number;
  matchesBeforeScheduledTime: number;
  surfaceBreakdown: HeadToHeadSurface[];
  recentMeetings: HeadToHeadMatch[];
}