# lib/api-zod/src/generated/types/headToHeadMatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';

export interface HeadToHeadMatch {
  id: number;
  date: Date;
  tournamentName: string;
  surface: string;
  playerA: Player;
  playerB: Player;
  /** @nullable */
  winnerId: number | null;
  /** @nullable */
  resultSummary: string | null;
}