# lib/api-zod/src/generated/types/headToHeadSurface.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HeadToHeadSurface {
  surface: string;
  totalMeetings: number;
  playerAWins: number;
  playerBWins: number;
  playerAWinPercentage: number;
  playerBWinPercentage: number;
}