# lib/api-zod/src/generated/types/matchDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HeadToHead } from './headToHead';
import type { Match } from './match';
import type { Tournament } from './tournament';

export type MatchDetail = Match & {
  tournament: Tournament;
  headToHead: HeadToHead;
};